% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAuxi.R
\name{input_popup}
\alias{input_popup}
\title{pop up window to choose fasta folder}
\usage{
input_popup(id, dir)
}
\arguments{
\item{id}{id}

\item{dir}{directory to monitor}
}
\description{
pop up window to choose fasta folder
}
