% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAuxi.R
\name{loadObj}
\alias{loadObj}
\title{check the folder, if no object, then process. If exist, then load directly}
\usage{
loadObj(f, ncores)
}
\arguments{
\item{f}{a folder containing the fasta files. Need read and write permission}

\item{ncores}{n cores passed to mclapply}
}
\value{
an object to be visualized
}
\description{
check the folder, if no object, then process. If exist, then load directly
}
