% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideSet.R
\name{peptideSets}
\alias{peptideSets}
\title{Get peptides set}
\usage{
peptideSets(files, mc.cores = 1, subset = NULL, ...)
}
\arguments{
\item{files}{a named character vector indicates the path of input fasta files. 
The names of the vector will be used in the figure. If the vector does not 
have name, name will be created from the file name.}

\item{mc.cores}{number of cores used to digest peptides}

\item{subset}{either a character vector of peptides or a \code{data.frame} of two columns. 
if a character vector is given, then the digested peptides will only be kept
if the peptides are present in the input vector. If a \code{data.frame} is given,
the first column should be a vector vector (see above) and the second column 
should be a numeric vector, which usually indicates the intensity of peptides
in an experiment. Then, the statistic is calculated not only from the presence 
or absence of peptide, also the intensity information will be used.}

\item{...}{other parameters passed to \code{trypsinDigestList}}
}
\description{
Get peptides set
}
