% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideSet.R
\name{peptideStats}
\alias{peptideStats}
\title{Stats peptides set}
\usage{
peptideStats(bmat, filter = FALSE)
}
\arguments{
\item{bmat}{binary matrix returned by \code{peptideSet} function}

\item{filter}{whether to filter out peptide whose weight < 0}
}
\description{
Stats peptides set
}
