% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{trypsinDigest}
\alias{trypsinDigest}
\title{In silico trypsin digestion of a protein}
\usage{
trypsinDigest(
  x,
  maxMissed = 2,
  minLen = 5,
  maxLen = 52,
  annot = NA,
  cleaveProline = FALSE
)
}
\arguments{
\item{x}{a character of peptide/protein sequence}

\item{maxMissed}{max miss cleavages allow}

\item{minLen}{the minumum length}

\item{maxLen}{the maximum length}

\item{annot}{user defined annotation for the protein}

\item{cleaveProline}{logical; whether proline should be cleaved}
}
\description{
In silico trypsin digestion of a protein
}
