% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{trypsinDigestList}
\alias{trypsinDigestList}
\title{In silico trypsin digestion of proteins in a list}
\usage{
trypsinDigestList(x, ..., mc.cores)
}
\arguments{
\item{x}{a list of proteins}

\item{...}{other arguments passed to trypsinDigest}

\item{mc.cores}{passed to parallel::mclapply}
}
\description{
In silico trypsin digestion of proteins in a list
}
